+++ want_to_do/src/states/todo/isAddTodoFormDisplayedState.ts
@@ -1,0 +1,6 @@
+import {atom} from "recoil";
+
+export const isAddTodoFormDisplayedState = atom({
+  key: 'isAddTodoFormDisplayedState',
+  default: false,
+})
