+++ want_to_do/src/states/todo/todoToUpdateState.ts
@@ -1,5 +1,5 @@
 import {atom} from "recoil";
-import {Todo} from "../hooks/todos";
+import {Todo} from "../../hooks/todos";
 
 export const todoToUpdateState = atom<Todo | null>({
   key: 'todoToUpdateState',
